Nearly added attack damage

+++ src/encounter.rs
@@ -14,5 +14,5 @@ pub struct Encounter {
 pub struct Enemy {
     pub name: &'static str,
     pub life: i16,
-    pub attack: u8,
+    pub attack: i16,
 }
