+++ src/player.rs
@@ -1,5 +1,5 @@
 pub struct Player {
     pub name: &'static str,
     pub life: i16,
-    pub attack: i8,
+    pub attack: i16,
 }
